Filter out doc comments in source sections
In a piece of documented code such as:
```
//! @struct buffer
//!     A buffer for holding some data.
//!     This is just an example so don't read into it too much.
struct buffer
{
    //! @member data
    //!     Contents of this buffer.
    char data[BUFFER_MAX_LENGTH];
    //! @member length
    //!     Length of the buffer in bytes.
    uint8_t length;
};
```
the information for members `data` and `length` are documented by their
@member tags.
Prior to this commit, these sections would "appear twice" in generated
documentation, once in plain HTML generated form the @member tags and
once again as raw source code when the struct body was printed.
This commit changes print_section to filter out doc lines as a means to
prevent this issue.